2.5 Сделал рефакторинг кода
